move print_scaling_variables() to use ScaleTensor
